requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.2.1
dm-tree==0.1.9
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.21
grpcio==1.74.0
h5py==3.14.0
huggingface-hub==0.34.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
k2==0.3.0
keras==2.14.0
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
onnx==1.16.0
onnx-tf==1.10.0
onnxruntime==1.18.0
opt_einsum==3.4.0
packaging==25.0
pluggy==1.6.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyreadline3==3.5.4
pytest==8.4.1
PyYAML==6.0.2
regex==2025.7.33
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scipy==1.16.1
sentencepiece==0.2.0
six==1.17.0
soundfile==0.13.1
speechbrain==1.0.3
sympy==1.14.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-addons==0.22.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.22.1
termcolor==3.1.0
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.54.1
typeguard==2.13.3
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
